State: Upgrade step to lower-case _id of envUser
Add an upgrade step (targeting 1.23) which lower cases the
part of the ":" envUserDoc "_id" field. Add an internal
function to state, docIDToLower, which does this. Ensure no more than
one envUser can be created with the same, case insensitive, name and
that state.EnvironmentUser() lookups are case insensitive.

Manually tested upgrade from 1.21 to 1.23.