storage: don't panic listing with multiple storage
"juju storage list" is panicking when there are
multiple storage instances/attachments for a unit.
The cause is the inappropriate use of
State.UnitStorageAttachments (was State.StorageAttachments)
for getting attachments for a storage instance.

This branch renames State.StorageAttachments to
State.UnitStorageAttachments, and adds a new
State.StorageAttachments that returns all storage
attachments for a specific storage instance.

I've also fixed improper storage status in one
case; there is more work to do there.