apiserver/storage: don't assume volume kind
When checking if the storage is persistent,
we no longer assume the storage is a block
kind with a volume backing it. We should
later move the code to compute storage info
into apiserver/common/storage.go.

We also no longer return ErrPerm except for
the initial user role check. Users are allowed
to see everything (until ACLs?), so obfuscating
the errors is just unhelpful.