Fix for LP-1436871
`type sNoMethods <type>` is used repeatedly in cmd/juju/status.go to
alias different types.

Although each type declaration is in it's own scope, which is acceptable
according to [the
spec][http://golang.org/ref/spec#Declarations_and_scope], gccgo finds
them as conflicting declarations.

This workaround is to give each one a unique name (except for
ssNoMethods, because both declarations are for the same underlying type,
and don't seem to bother gccgo).