all: fix declared by not used error in tests
gccgo is stricter than gc in this respect and discovered some unused
variables which were obviously supposed to be hooked up to tests.

This approach isn't actually thread safe, but as the rest of the tests
in this suite also commit the same crimes, it feels like the lesser evil
to fix the test and unblock gccgo.

Change-Id: Id49cc6d1c289a1cbf170f03acd0488774f70a702