Fix up a bunch of tests and make tracker shutdown cleanly.
It was using errors.Trace() to give nicer feedback wrt failures. However, this ends
up masking errors that we compare directly (tomb.ErrDying being the one that can happen
in tests).

Unfortunately we now have a non-deterministic test. Specifically when we set the current
unit to dying, we will get both a 'relation-departed' and 'relation-broken' but also
a 'leader-settings-changed' to inform us we're no longer the leader.
However, leader-settings-changed occurs asynchronously from the others, so we can't
rely on an exact firing time. I don't know how to fix this yet.