Fix govmomi imports.
This is necessary because the github.com/vmware/govmomi/vim25/soap
package has code that does works under Go 1.3, and juju must support
Go 1.2 still.  So we have forked the govmomi repo and patched it to work
under Go 1.2.  This requires that all the imports in the vmware provider
be changed to the fork.  Once juju's min Go version is 1.3 or higher, we
can switch it back.