Fix lp:1446264 by using joyent kvm packages only
Currently the joyent provider does not distinguish between
kvm and smart package types, it just assumes all listed
packages are kvm. This causes CreateMachine to fail if it
happens to try combining a smart package with a kvm image.

Unfortunately, the Joyent ListPackages api does not include
the virt type, however according to the Joyent developers
only the kvm packages have VCPUs set to one or greater.

Change the joyent provider to correctly label the packages
virt type based on the number of vcpus, and constrain
InstanceSpec selection to kvm packages and images only.