Replace windows services package
This PR branch replaces the current windows service package, which is highly dependent on executing powershell commandlets with a native implementation.

For the time being, the securestring package will be added to the windows service package. It has to be made more generally useful before we can split it out into its own project. It still lacks entropy, encrypting binary blobs and using encryption stores other then the user one.