WIP - Add syscmd base and system environments cmd
This patch is for the jes-cli feature branch and is a
work in progress.  It introduces:
1 - A base for system related commands which enables
those commands to lookup and create API connections
for the current system.  Currently, we are using the
current environment name as the system name until
work is done to implement a current-system file.
Tests for this command base will be written as part
of that work.
2 - The system environments command.  This command
will list out the environments on the current server
which the current user has access to.