Allow server only writes to cachefile
This patch updates logic in the cachefile
config store to allow server-only writes.
This can happen when a user uses the login
command for a server where they will not
be placed in an environment.

This patch also consolidates the JES and
EnvironmentsCacheFile feature flag as the
work done under the JES flag will depend
on the cache file.