provider/openstack: wait for volume provisioning
When we issue a CreateVolume to Cinder, Cinder will
(or may?) return incomplete information about the
volume, lacking status or size, for example. We must
wait for the volume to transition to or past "creating",
so it has a size that we can record in state.

Fixes https://bugs.launchpad.net/bugs/1450740