state: prevent incompatible charm storage upgrade
When a service upgrade is requested, we will now
check the differences in storage between the current
and new charm revisions. We will prevent upgrades
under any of the following conditions:
 - new required storage is added
 - range of existing storage is contracted
 - type, shared, read-only or location of existing
   storage changes

It is therefore acceptable only to add new storage
if it is optional, and existing storage ranges can
only be increased (min decreased/max increased).
Storage description and properties are not considered.