worker/storageprovisioner: watch machines
When a volume or filesystem is to be attached
to a machine, and the machine is unprovisioned,
then the storageprovisioner worker will now
watch that machine. When the machine is observed
to have been provisioned, we will update the
parameters of the associated storage entities
with the instance ID and proceed to create them.