CMD/JUJUD: Scope Env Storage Provisioner for Hosted environ
Storage volumes in hosted environments were unable to be watched,
because the scope of the storage worker was being set to that of the
state server instead of the hosted environment. This then lead to the
StorageProvisionerAPI ScopeAuthFunc denying the privilege to watch the
entity as the environTag passed in (erroneously state server) != the
hosted environTag.

Fix: Set scope to that of hosted environment in machine.go and add
test coverage.