apiserver: fix data race in tests
This PR removes the data race on the timer reference captured using PatchValue.
timer's cannot be reset across multiple goroutines, the only part that is safe
to use concurrently is the timer.C channel.

Resetting the timer was only used during the latter part of the test to avoid waiting
for the entire 3 minutes, so replace that with a call to set the timer to an initally
small value, then adjust the test to work around multiples of this value.