provider/ec2: tag instances and volumes
We tag instances started with StartInstances
with the machine tag and the environment UUID.
Similarly, we tag volumes with volume tag and
environment UUID.

State server instances are additionally tagged
with a "JujuStateServer" tag. This is not used
for anything yet, but we can later use it to
improve the way we identify state servers.

In the console, instances and volumes will now
display the tag as the instance's "name" column.
The environment UUID can be used for identifying
instances and volumes within an environment, e.g.
for AllInstances, and also for billing purposes.