Fix for 1442308: fix cloud-config for systemd
Juju was previously unable to create vivid containers
due to problems with the service started with systemd
to shutdown the template container.  The problems
addressed were:

- Specifying an "After" for systemd should not generate
a "Conflicts" entry.  Starting the service would cause
the conflicting service to be killed.  The proper way
to specify that a service should wait for another service
to complete before starting is to use a target.  For our
case, the target to specify that cloud-init has finished
is cloud-config.target.

- Even transient services must include an Install section.
This is required for services registered with systemctl.

- The service must be explicitly started for systemd.