Force the DB logging to be on if JES is enabled
If the JES feature flag is enabled then the we want it to be as if the
DbLog feature flag was enabled too. A helper function is introduced to
help with his and the DbLog constant is no longer exported to enforce
usage of the helper (tests now use the feature flag name directly
instead of the constant).