cmd/juju: fix #1447899 upgrade juju fails
When upgrading an agent from a client that is more than one minor
version ahead, the agent's minor version is incremented. This branch
resets the agent's original patch version to zero. If this is not
done, juju will not find tools if the new version has no patch larger
than the old version's patch.

For example, upgrading a 1.20.14 agent using a 1.22.1 client will
result in a desired version of 1.21.14 - but if the newest available
1.21 tools is 1.21.3, juju will complain that no compatible tools are
available. This branch corrects the bug by returning, in this
scenario, a desired version of 1.21.0 and thus finding the 1.21.3
tools.