Fix 1459093: Upgrade fails if there's an unknown series in streams
Allow tools metadata with an unknown OS to be merged into
simplestreams. Such tools will not have a known series version.
Introduce an "unknown" series version. Use this to construct the
metadata's productId under which these tools are catalogued in
the simplestreams JSON output.

In addition to testing the above, test that upgrade-juju succeeds with
an unknown series in simplestreams.