mongo: fix oplog tests on precise
oplogSuite was based on IsolationSuite which wipes out $PATH. This was
preventing MgoInstance from finding the installed mongod on precise
systems. BaseSuite is now used instead.

Also fixed a test issue with TestDiesOnFatalError when mongodb 2.6 is
used. If the oplog collection was empty the tailer was created it
would would not stop when the oplog collection was dropped.