worker/storageprovisioner: react to Dead vo/fs
When a volume or filesystem becomes Dying, we
now simply clear out any pending provisioning.
When a volume or filesystem becomes Dead, then
we destroy (if provisioned) and remove from
state. The entity being Dead is predicated on
it having no dependents; this will be taken
care of in state.

Note: currently we don't have a DestroyFilesystems
method in storage.FilesystemSource.