apiserver: fix LP 1466011
Fixes LP 1466011

http://reviews.vapour.ws/r/1928/ removed a bunch of string munging logic related to the reported address of the api server.

This change exposed an expectation by the apiserver test cases that the apiserver would _always_ report it was listening on "localhost", even though the test suite never bound the socket to localhost in the first place. Fix this by explicitly binding to localhost as the test expects, and in the process remove a bunch of code in the api client which had grown to expect a set of 'special' addresses in the api.Info.Addrs slice.

As _real_ clients always use the address specified in their jenv file, that is they never are asked to connect to "0.0.0.0" or "::" or "localhost", the behaviour of the test now better mirrors how the code is used.