Pass stdin to jujuc if prompted
The jujuc CLI will now pass through stdin to
the server if prompted with a special error.
If the client has not provided stdin, the
server passes a special io.Reader to the command
that will yield the special error; this is
sent to the client, which then re-executes
the command after consuming os.Stdin.

Fixes https://bugs.launchpad.net/juju-core/+bug/1454678