api: fix data race in Client.AddLocalCharm
Updates LP 1465115

A data race exists because the Client holds a reference to the charm bundle file descriptor and closes it unconditionally on the way out of the method.
However, the net/http Client also has a reference to the file descriptor via the http request body and also closes it if the upload is successful.
It is not clear if the body will be closed if the request is not successful.

This creates a race on the Close method which is not safe to be called by multiple goroutines. Work around this by inserting a wrapper which nullifies the Close method presented to the http Client.