state: add Binding() to Volume and Filesystem
We add the Binding() method to Volume and Filesystem,
backed by a new "binding" document field. This field
identifies an entity whose destruction triggers the
destruction of the volume or filesystem.

This is the mechanism we will use to ensure that, by
default, volumes and filesystems will be destroyed
when their associated storage is destroyed. Volumes
that back filesystems are bound to the filesystem;
volumes and filesystems directly related to storage
instances will be bound to said storage; otherwise
volumes and filesystems may be bound to the environment
(destroy when the environment is (cleanly) destroyed),
to the machine they're attached to (destroy on detach),
or to nothing (don't destroy ever, unless explicitly
requested).