storage: loop creates Persistent volumes in LXC
The loop storage provider will now mark volumes
as Persistent when used within an LXC container,
including the local provider. This is necessary
to ensure that containers are not destroyed until
the loop devices are detached, since they will
outlive the container.

A new lxcutil package has been introduced, with
a single function extracted from container/lxc.
This is necessary to prevent an import cycle.