apiserver/storageprovisioner: +RemoveAttachment
Implement the RemoveAttachment method in the
apiserver/storageprovisioner API server facade.
The client side is in place, and is hooked up
in the worker; it just hasn't been triggered
due to missing lifecycle support, which will
be landing soon.

Also, some related changes to
Remove{Volume,Filesystem}Attachment in state,
to return a NotFound error if the attachment
isn't found initially.