worker: fix data race in FakeState
Fixes LP 1468188

After a discussion it was decided to make the FakeState's set/get EnvironConfig methods look more like the _real_ state, which backs the config.Config value with a map stored in mongodb, rather than sharing a copy of *config.Config.

This has the side effect that each value passed or returned from SetConfig, EnvironConfig will produce a new *config.Config value which is not shared by other callers.