state: prevent Machine.EnsureDead with storage
Machine.EnsureDead will now fail if the machine
has non-inherently-machine-bound storage associated
with it. This includes persistent storage, and any
unprovisioned dynamic storage, since we don't know
whether or not it will be persistent once
provisioned, and that could happen concurrently.

If such storage is present, then EnsureDead will
fail with an error that satisfies
state.IsHasAttachmentsError.