worker/storageprovisioner: order detach correctly
Detaching filesystems requires that the we store
information about the filesystem first, since we
need to know whether or not the filesystem is
volume-backed, for example. So we change filesystem
detachment to not be immediate, and use the state-
machine/pending style used before, with detachment
depending on filesystem info being seen.