state: use supported type when writing leadership settings docs
A bare empty map is written to the settings collection by
addLeadershipSettingsOp which is unsupported by the multi-env txn
layer. This means that the document is written out without the
env-uuid field.

This is the bare minimum fix. There be further PRs which improve the
multi-env txn layer to avoid this kind of problem in the future.

This is part of the fix for LP #1468994.