JES Deploy Test
Test that Juju can create two hosted environments and deploy the dummy stack
in each, as well as the system environment. Add unit tests. Ensure cache.yaml
and .jenv files are handled correctly and that the client's shell environ can
be initiated with the jes flag. All unit tests and manual testing  pass.