worker: always clear workerInfo.worker before restarting a worker
If a worker is being restarted, make sure that the previous Worker
reference is removed before starting a new one.

Rather than the previous situation where the Kill signal could be delivered
to a previous copy of the worker, this change will cause the runner to
continue to try to issue killAll, which will do nothing as there are no non nil
workerInfo.worker's registered in the worker map, until the new worker reference
is recieved via runner.startedc. At this point the killAll loop will then be able
to issue the Kill command to the correct copy of the worker.