state: improve env UUID migration robustness
Handle docs that already have_ids with the env UUID prefix but don't
have the env-uuid field. This was happening b/c the leadership code
was active during upgrades and for complicated reasons was writing out
documents with correct _id fields but the without the env-uuid field.

See LP #1468994 for more details.