state: <location>/<subdir> for multiple storage
If a charm storage can have multiple instances,
and it specifies a location, then we should be
creating a sub-directory for each storage
instance. If the storage is a singleton, then
we should use the location as-is.

At the same time, we reject storage that
specifies a location inside the Juju storage
directory, which is reserved for Juju.