State: Introduce an Environment Count
Introduce an environment count that is incremented / decremented
each time an environment is respectively created / destroyed. The
system environment is not counted.

This is a prerequisite to refactoring environment Destroy. A future
branch will use the count to ensure the system environment cannot be
deleted if the environment count is > 0.