apiserver: fix data race in certChangeListener
Fixes LP 1467753

This PR takes the cl.m lock during certChangeListener.Accept().
This ensures that the cert change listener cannot change the TLS config
concurrently with accepting a new connection.

This does have the downside that because Accept is a blocking call, the lock
will be held until an incoming connection is recieved. This means the cert change
requires at least one incoming connection to be _accepted_ (not completed) before
taking effect.

In practice this does not appear to be an issue.