apiserver: fix the data race for tls.Config
Each connection gets a private copy of the tls.Config so that only
future connections are affected by certificate updates. This avoids
the race and is just generally safer.

Also simplify the changeCertListener somewhat.

This also fixes a serious bug recently introduced in c147767d. The
addition of the Lock call at the top of processCertChanges caused the
cert update to hang when updateCertificate tries to grab the lock
again.