Forward Port: fix bug in namespace usage (lp:1461354) #2489
fixes-1461354

https://bugs.launchpad.net/juju-core/+bug/1461354
debug-log EOF: invalid tag and panic seen in state server logs

This was a result of me changing the namespace stored in the config to no longer store the leading -. Instead, it just stores the namespace itself, and where we use it, we have to format it correctly. This seems like a much more sane way to do it, so the value you set is the value that exists, and we don't dictate formatting to everyone using the namespace.... however, it meant I forgot to check for an empty namespace in this one case.