upgrades: migrate 1.24.{0,1} rsyslog config
In 1.24.0, there was an upgrade step introduced
that migrated rsyslog certs and other config files
from LogDir to DataDir. In 1.24.2 this was changed
to go from LogDir to ConfDir, but there was no
consideration for environments that were already
upgraded to 1.24.0.

This means the rsyslog config gets left behind in
DataDir, and new config gets generated by the rsyslog
worker. Each state server does this, which then
causes an initial problems with inconsistencies
between CA certs and certs generate for rsyslog.

We fix this by checking in both DataDir and LogDir
for config to migrate; we stop at DataDir if we
find it there, otherwise check in LogDir.

Fixes https://bugs.launchpad.net/juju-core/+bug/1474614