cmd/jujud/agent: detach loop devices on uninstall
When uninstalling the machine agent, if running in
LXC then prior to removing the data-dir we will now
detach loop devices on the machine. This is necessary
to prevent loop devices leaking.

The container destruction code will also do this,
but since the data-dir is removed here we must do
it in both locations or we have a race between the
uninstall code removing data-dir and the container
destruction code checking data-dir for loop device
backing files.