state: add env UUID to multiwatcher.EntityId
To allow a single multiwatcherStore to track entities for multiple
environments, EntityId include the environment UUID for the
entity. This is needed to support a new watcher (coming soon) which
will watch all environments.

Also in this change: extracted the logic for State.docID and
State.localID into standalone functions so they can be used
independently (there will be more uses of these soon).