Improved status-history by removing sequence.
Status history used a sequence as _id this was a poor decision
because it caused excessive writes to a given doc and also produced
a bug (lp:1479289) where statuses with clashing data where inserted
from various envs.
I changed the _ids to be handled by mgo, changed entityid to be
named globalkey as it is what it actually is and changed updated
field to hold an int64 thus improving precission of the stored time.