provider/ec2: don't attach volumes initially
Change CreateVolumes to not attach volumes initially,
but instead defer to AttachVolumes. This allows the
volume to stay in "creating" arbitrarily long without
failing the CreateVolumes call; AttachVolumes will
then be retried until the volume is in the "available"
state.

Also, fix a bug in DescribeVolumes to do with
identifying whether or not a volume is persistent.

Fixes https://bugs.launchpad.net/juju-core/+bug/1479546