state: volumes shouldn't prevent destroy-env
Some time ago we added code to prevent destroy-environment
from succeeding if any persistent volumes were present in
the environment. This was necessary to prevent (except
through use of --force) leaking persistent volumes.

This is no longer necessary, as we now destroy all volumes
in the environment via env.Destroy().