state: more robust environment death handling in allEnvWatcher
Changed the signature of removed() on the backing structs so that it
takes the env UUID separately from the *State. The *State is now
allowed to be nil.

This allows removal of entities from the store when an entity's
environment as been removed. When allEnvWatcher.Changed() fails to get
a State for an environment it now checks if the environment is still
alive. If it isn't, the entity is removed from the store.