worker/storageprovisioner: set status
After attempting to create or attach a volume, successfully or
otherwise, we now set the status of the volume. If volume
creation fails, we keep the status as "pending" but set the
status message to the error; this provides some feedback, but
also indicates that it's not a fatal error and is still pending.
Similar for "attaching".

We will expand this later to handle detaching and destroying.