mongo: EnsureServer now always write out mongodb related config files
We had seen reports of the mongod init system config being in place
but the shared secret file not being there. Although it is not
understood exactly how this happens, ensuring that the shared secret
file, SSL key file and monogdb config file are all in place
irrespective of the init system config is more robust and should fix
the reported problem.

Fixes LP #1468579.