state: return StatusData consistently from All(Env)Watcher
When an entity's StatusData was empty it would sometimes be returned
as nil and other times as an empty map. This makes the tests hard to
write and was causing intermittent test failures. It also makes
handling the output of the WatchAll and WatchAllEnvs API calls more
difficult for a caller (they had to handle 2 types).

A lack of status data is now always returned as an empty map.