LP 1486749: Disallow backups for hosted envs
This patch disallows any backups command for a hosted
environment, as it is currently not supported. It prevents
an API connection to the Backups endpoint if not operating
on the system (state server) environment.

The doc for juju backups was updated to indicate that backups
are not supported for hosted environments, but only if the
JES flag is enabled, to avoid confusion for those who are not
running with this feature yet.