container/{kvm,lxc}: use arch.HostArch in place of version.Current.Arch
This PR is a small part of a set that pull apart the confusing mess that
is version.Current.

This PR replaces several uses of version.Current.Arch, and subsiquent
monkey patching of the value with arch.HostArch. In this instance their
usage is synonumous so there is functional change (apart from breaking
the requirement to import version)