Update status test passes
Added status update operation to the resolver loop - if there are more operations of this kind
(triggered by a channel, but independent of the remote state, we could do it in a smarter way,
but for now this should do).

As a drive-by created config structs for the remotestate/watcher and the resolveloop, which were taking
large-ish numbers of parameters.