all: remove reference to version.Current.Arch
Remove all remaining references to version.Current.Arch and replace
with arch.HostArch(). For tests, replace the version.Current.Arch
patch with a patch of arch.HostArch.

This PR enbles removal of version.Current.Arch and makes it clearer
which code paths explicitly depend on the architecture of the _host_
machine.