worker/provisioner: Remove indirection from FindTools
FindTools is threaded through the api and the provisioner. Some
of these methods take the arch as an optional string (blank is
accepted to be "choose the default"), and others as a pointer to a
string.

This difference is not useful as it requires more effort to call
FindTools at some points in the process. To make this simpler,
replace the *string in some signatures with a string value, which
simplifies the internal API, the callers of which already had a
string value.