worker/uniter: handle conflicted upgrades
This PR fixes conflicted upgrade handling. If an
upgrade fails due to a conflict, we now set the
LocalState's Conflicted field to true, and then
we clear it when an upgrade operation is committed.

When in the conflicted state, we only accept two
methods of progressing: hook resolution, or a
forced upgrade. One creates a resolved-upgrade op,
and the other creates a revert-upgrade op. If
neither of these is true, we'll just for a state
change.

We also simplify handling of the Stopped local
state. If the unit has already run the Stop hook,
then it should not generate any more operations.
So we now check for life==Dead or Stopped at the
top of NextOp.