worker/uniter: fix various resolver bugs
leadership: don't run hooks unless we're in the Continue
state, and not Stopped. Leave queued leader-elected hooks
to the top-level resolver.

relations: don't run hooks unless we're in the Continue
state, and not Stopped.

storage: don't run hooks unless we're in the Continue
state, and not Stopped; or if we're not yet installed,
and there's a queued install hook. Don't issue storage-
attached hooks for Alive but unprovisioned attachments.

uniter: don't run install hook if the uniter is Dying.

tests: eliminate error log message from upgrade test.
Update TestUniterConfigChangedHook to not expect a
config-changed after resolving a failed config-changed.