worker: new undertaker worker
Introducing The Undertaker (http://goo.gl/ZGxqNB): a new periodic
worker which is started for an environment when that environment is
set to dying. The undertaker waits for all machines and services to be
removed from the environment, checking every 5 minutes. Then, it sets
the environment's life to dead and stops.

The envWorkerManager, who is watching the environ's life-cycle, then
sees that the environment is dead and starts a goroutine which removes
all environment docs 24 hours from time of death.

The resumeUndertaker method on the envWorkerManager ensures that, in
the case of an unexpected reboot or the like, the undertaker picks up
where it left off.