apiserver: DestroyEnvironment
DestroyEnvironment force destroys machines and services and does not
call RemoveAllEnvironDocs directly. Once the provisioner has
decommissioned all resources and no machines or services are left in
the environment, envWorkerManager handles the removal of the
environment.

As environment removal is now asynchronous, several tests, which
expect the environment to be immediately removed, have been skipped
(in particular the DestroySystem tests). This allows this branch to be
landed in the target feature branch. A follow up branch will make
DestroySystem wait for environment removal to complete.