version: remove version.Binary.OS
version.Binary.OS is only used in two contexts, the first as the valu
version.Current.OS. The second for tools sorting and bootstrapping.

In the first case version.Current.OS will be identical to os.HostOS(),
so replace all the calls to the former with the latter.

In the second case, version.Binary.OS is _always_ deducable from
version.Binary.Series using version.GetOSFromSeries, so do that, then
we can drop the field altogether.