storage: remove persistent as pool configuration
It no longer makes sense to have "persistent" as
a pool configuration attribute. Instead, providers
create persistent storage if they support it, or
otherwise don't (there *could* be provider-specific
config for this, but so far it has been unnecessary).

Later we'll have the ability to manage the lifecycle
binding of storage, which will cause persistent
storage to live beyond attachments.