Refactoring cmd/juju
Working on the "Chicago Cubs" branch, we found a few problems with out CLI commands.

After a discussion with rogpeppe, jam and fwereade, it was decided, that our command unit
tests should be testing s "wrapped" commands (using either the environment or system command
wrapper).

To enable macaroon based login mechanism in the "Chicago Cubs" branch, those wrappers were
used to initialize httpbakery Client (including loading macaroons from a persistent cookiejar)
and each command that needs an API connection will need to be wrapped, as wrapped initialize
an "apiContext" structure, which can then be used to initialize an API connection.

This is the first part of changes moving slowly towards that goal.