api/storage: StorageDetails to match model
This is the first in a series of changes to
fix the "storage" API facade and storage CLI
commands to better reflect the model. This
is necessary to cater for future changes,
to support shared storage.

This change in particular renames the existing
apiserver/params.StorageDetails to LegacyStorageDetails,
and introduces a new StorageDetails. On the
server we always populate both legacy and new.
A follow-up will change the CLI to use the new
details structure, falling back to legacy if
the new one is not available (i.e. old server).