simplestreams: Ensure HTTP resp.Body will always get closed
In some cases (explicit StatusCode handling) the HTTP/GET resp.Body in
Fetch() does not always get closed leading to file descriptor leaks.

It is unfortunate that we do the opening in Fetch() and the Close()
somewhere else (i.e., the caller). It would be much better if we could
rely on a single `defer resp.Body.Close()'.

To fix this I'm ensuring we close the body in those particular cases
where we introspect StatusCode and choose to return an error instead of
the response.

The documentation for http.client.Get() explicitly calls this out:

  // Caller should close resp.Body when done ...

Fixes [LP:#1496750](https://bugs.launchpad.net/juju-core/+bug/1496750)