apiserver: rationalize error handling
We derive HTTP status codes from errors and consolidate as much of the
error response handling as we can.

The only semantic change in this branch is that the CodeOperationBlocked
error now returns an http.StatusForbidden status code rather than
StatusBadRequest.  Given that http status codes are ignored other than
for success/failure, this should have no impact.

In passing, we also change common.ErrOperationBlocked to OperationBlockedError,
because that is the idiomatic form for functions that return new errors,
and its error string to all lower case, because that is also standard.