storage: update filesystem API/CLI
This commit changs the FilesystemDetails structure
to be like VolumeDetails, embedding a StorageDetails
with the storage and unit-attachment information.
This is then used in the updated CLI to display
attachment information in a way that will work with
shared storage in the future.

Primary difference to volume listing is that there
is no legacy structure for filesystem details, as
filesystem listing does not exist in 1.25 or earlier.