apiserver/common: fix StorageAttachmentInfo
StorageAttachmentInfo was failing for block-kind
storage created by the Azure provider. This is
due to the fact that the only information we have
to identify disks is the SCSI LUN/bus address.
Rather than relying on the volume/attachment info
alone to identify the storage location, we find
the matching block device and return its path.

Fixes https://bugs.launchpad.net/juju-core/+bug/1498746

Conflicts:
	apiserver/storage/package_test.go