APIServer: DestroyEnvironment
DestroyEnvironment schedules machines and services in the environment
to be destroyed. It  does not RemoveAllEnvironDocs directly. Once the
provisioner has decommissioned all resources and no machines or
services are left in the environment, the undertaker worker handles
the removal of the environment.

As environment removal is now asynchronous, several tests, which
expect the environment docs to be immediately removed, have been
skipped (in particular the DestroySystem tests). A follow up branch
will make DestroySystem wait for environment removal to complete.