apiserver: fix data race in shutdown
Fixes LP 1500283.

Fixes data race in any test using mongo/state which was introduced in 5cec732d.

To make sure the reference to mongoUnavailable is visible to all threads we need
to use an atomic operation (or a lock). There is no operation that acts on a bool
in the standard library, so we use atomic.LoadUint32.