apiserver/common: handle DischargeRequiredError in ServerError
We return an http.StatusUnauthorized error in this case.

We needed to move DischargeRequiredError from apiserver/authentication to avoid an import cycle.

Also clean up various places where we used a hack to get around
the immutable identity-url attribute to make macaroon-related
tests and use the new JujuConnSuite functionality instead.