provider/gce: set HardwareId, not DeviceName
The gce storage provider was setting VolumeAttachmentInfo.DeviceName
to an invalid value. When we create volumes, we specify a device
name to include in the /dev/disk/by/id path. This is stable across
attachments, so we can use that in the VolumeInfo.HardwareId field.

Fixes https://bugs.launchpad.net/juju-core/+bug/1500703