provider/ec2: don't use numbers in device names
When assigning device names to block devices,
don't use number suffixes. Doing this confuses
some applications, such as Ceph. This reduces
the number of volumes we have available, so
if it is of value we can introduce an pool
configuration to enable adding suffixes.

Fixes https://bugs.launchpad.net/juju-core/+bug/1500721