storage: add BlockDevice.DeviceLinks
Add DeviceLinks to the BlockDevice and volume attachment
info models. The DeviceLinks field captures the OS (udev)
maintained symlinks to the block device, which a volume
provider can match on.

Update the GCE storage provider to populate the DeviceLink
field instead of HardwareId, adhering to the documented
method of identifying disks.

https://bugs.launchpad.net/juju-core/+bug/1500803

Conflicts:
	apiserver/common/storage_test.go
	apiserver/common/volumes_test.go