WatchStorageAttachment watches block devices
The unit agent needs to know when storage attachments
are provisioned and ready for use. A block-kind storage
attachment is not available until there is a block device
visible on the machine, so the uniter needs to be informed
of changes to block devices on the machine.

Fixes https://bugs.launchpad.net/juju-core/+bug/1501173