upgrades: introduce environ config upgrades
Introduce a new interface, environs.EnvironConfigUpgrader,
which an EnvironProvider can implement to provide it an
opportunity to alter environment config on upgrade. This
is currently only run on upgrade to 1.25, but the intention
is to have it run on any version upgrade when we need to
use it again. This should replace any use of Validate to
upgrade config.

Provider registration has been extracted into another new
interface, environs.ProviderRegistry, with the existing
functions redirected to a global provider registry object.
This was done to simplify testing of code that involves
environment provider lookup.

The GCE provider is modified to set the default block
source on bootstrap, and also on upgrade.

Fixes https://bugs.launchpad.net/juju-core/+bug/1500769