api: consolidate client websocket implementation
The protocol used by logsink and log is the same,
so make the code the same too. We expose a
new abstraction inside the api package, Stream,
which represents a streaming websocket connection
that uses the "first-line-is-error" connection protocol.

We implement an api/logsender package to
hide the low level API protocol details from the worker/logsender
package.

Now that we have the websocket protocols consolidated,
we will be able to implement macaroon authentication
in one place.