Always create the ubuntu user from cloud-config
Juju expects any ubuntu image to have either an ubuntu user already
created or ubuntu as the default in /etc/cloud/cloud.cfg for
cloud-init to pick up and use.

This change adds cloud-config rules to always create an ubuntu user
with sensible defaults instead of the default user configured in the
image. This enables bootstrap on Dreamhost, and as a side effect
cleans up the centos user creation to share the same configuration.