state: Add Environment Mode
Add a new "mode" field to the environment doc. Add SetMode and Mode
methods. Add two modes "normal" (default) and "frozen". Add a 1.26
upgrade step to add the mode field to all environment docs and set the
value to "normal". Change checkEnvLife to checkEnvLifeAndMode. This
ensures that the environment is normal as well as alive.