cmd/juju/status: added environment-status section (1.25)
1. The recently added top-level available-version section is now
underneath a new environment-status section. This section will later
be used to also show information about environment migration and other
high-level concerns. The environment-status section is only displayed
when there's something to be shown in it.

2. The display of an available upgrade was removed from the oneline
formatter. The upgrade notification was being shown as if it was a
unit which is confusing and could break automated parsing. Given the
tight focus of the oneline output (showing the units and subordinates)
I think it's best not to even attempt to include the upgrade
notification there.

3. The tests have been cleaned up. Instead of making it appear an
upgrade is available during all status tests - requiring that all
expected output include the upgrade notification, even when a test
case has nothing to do with the upgrade notification - there are now
specific tests for the upgrade notification.